+++ CHANGELOG.md
@@ -17,6 +17,7 @@ N/A
 * Improvements to maintainability regarding string-resources.
 * Added custom style for bottom navigation bar.
 * Replaced buttons to save and delete custom password quality gates with image buttons.
+* Added `CtActivity` and `CtContainer` class for easier integration of collapsing toolbars with image buttons.
 
 <br/>